# frontend/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/pomodoro/history": {
      "get": {
        "operationId": "pomodoroHistoryList",
        "parameters": [
          {
            "name": "configurationId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "COMPLETED",
                "ENDED_EARLY",
                "CANCELLED"
              ]
            }
          },
          {
            "name": "dateFrom",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "dateTo",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pomodoro history returned.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "object",
                          "required": [
                            "items",
                            "meta"
                          ],
                          "properties": {
                            "items": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/PomodoroHistoryResponseDto"
                              }
                            },
                            "meta": {
                              "type": "object",
                              "required": [
                                "page",
                                "pageSize",
                                "totalItems",
                                "totalPages"
                              ],
                              "properties": {
                                "page": {
                                  "type": "integer",
                                  "minimum": 1
                                },
                                "pageSize": {
                                  "type": "integer",
                                  "minimum": 1
                                },
                                "totalItems": {
                                  "type": "integer",
                                  "minimum": 0
                                },
                                "totalPages": {
                                  "type": "integer",
                                  "minimum": 0
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "The request is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected server error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List my Pomodoro phase history",
        "tags": [
          "Pomodoro"
        ]
      },
      "post": {
        "operationId": "pomodoroHistoryCreate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePomodoroHistoryRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Pomodoro history recorded.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/PomodoroHistoryResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "The request is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected server error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Record a completed or ended phase",
        "tags": [
          "Pomodoro"
        ]
      }
    },
...
  },
  "info": {
    "title": "Pomodoro and Music API",
    "description": "Backend API for Pomodoro sessions and music playback.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ApiErrorDetailDto": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string",
            "example": "name"
          },
          "message": {
            "type": "string",
            "example": "Name is required"
          }
        },
        "required": [
          "message"
        ]
      },
      "ApiErrorResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "example": 400
          },
          "message": {
            "type": "string",
            "example": "Invalid request"
          },
          "error_code": {
            "type": "string",
            "example": "INVALID_INPUT"
          },
          "details": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiErrorDetailDto"
            }
          },
          "request_id": {
            "type": "string",
            "example": "0198bb70-92ac-77c8-b22f-892d2d97f599"
          }
        },
        "required": [
          "code",
          "message",
          "error_code",
          "details",
          "request_id"
        ]
      },
      "ApiResponseDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "object",
            "example": "success"
          },
          "code": {
            "type": "number",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "Success"
          },
          "data": {
            "type": "object"
          }
        },
        "required": [
          "status",
          "code",
          "message",
          "data"
        ]
      },
...
      "PomodoroPhaseType": {
        "type": "string",
        "enum": [
          "FOCUS",
          "SHORT_BREAK",
          "LONG_BREAK"
        ]
      },
      "PomodoroHistoryStatus": {
        "type": "string",
        "enum": [
          "COMPLETED",
          "ENDED_EARLY",
          "CANCELLED"
        ]
      },
      "PomodoroHistoryResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "pomodoroId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "configurationName": {
            "type": "string",
            "nullable": true
          },
          "phaseType": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PomodoroPhaseType"
              }
            ]
          },
          "plannedDurationSeconds": {
            "type": "number"
          },
          "actualDurationSeconds": {
            "type": "number"
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PomodoroHistoryStatus"
              }
            ]
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "endedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "pomodoroId",
          "configurationName",
          "phaseType",
          "plannedDurationSeconds",
          "actualDurationSeconds",
          "status",
          "startedAt",
          "endedAt"
        ]
      },
      "CreatePomodoroHistoryRequestDto": {
        "type": "object",
        "properties": {
          "pomodoroId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "phaseType": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PomodoroPhaseType"
              }
            ]
          },
          "plannedDurationSeconds": {
            "type": "number",
            "example": 1500
          },
          "actualDurationSeconds": {
            "type": "number",
            "example": 1500
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PomodoroHistoryStatus"
              }
            ]
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "endedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "phaseType",
          "plannedDurationSeconds",
          "actualDurationSeconds",
          "status",
          "startedAt",
          "endedAt"
        ]
      },
...
    }
  }
}